Add abstract Socket
